+++ Taller Corralon/src/corralon/vistas/materiales/detalleMateriales.java
@@ -61,11 +61,11 @@ public class detalleMateriales extends javax.swing.JPanel {
        //material.setIdMaterial((Long)id.getValu);
        material.setCodMaterial((Long)codigo.getValue());
        material.setNombreMaterial(nombre.getText());
-       material.setPrecio((float) precio.getValue());
+       material.setPrecio((Long) precio.getValue());
        material.setDescripcionMaterial(descripcion.getText());
-       material.setCantidadExistente((int)cantidad.getValue());
-       material.setStockMaximo((int)stockmax.getValue());
-       material.setStockMinimo((int)stockmin.getValue());
+       material.setCantidadExistente((Long)cantidad.getValue());
+       material.setStockMaximo((Long)stockmax.getValue());
+       material.setStockMinimo((Long)stockmin.getValue());
     }
  
     public detalleMateriales() {
